requirements.txt
django>=5.1,<6.1
gunicorn>=21.2
whitenoise>=6.7
psycopg2-binary>=2.9
djangorestframework>=3.14
django-cors-headers>=4.3
requests>=2.32
Pillow>=10.4
python-dotenv>=1.0
